fix sourcemaps and cleanup configs
