+++ .make-packages.js
@@ -3,7 +3,6 @@
 const pkg = require('./package.json');
 const fs = require('fs-extra');
 const mkdirp = require('mkdirp');
-const { copySources, cleanSourceMapRoot } = require('./.make-helpers');
 
 const ROOT = 'dist';
 const CJS_ROOT = `${ROOT}/cjs`;
@@ -14,13 +13,24 @@ const PKG_ROOT = `${ROOT}/package`;
 const CJS_PKG = PKG_ROOT;
 const ESM5_PKG = `${PKG_ROOT}/esm5`;
 const ESM2015_PKG = `${PKG_ROOT}/esm2015`;
-const SRC_ROOT_PKG = `${PKG_ROOT}/src`;
 const TYPE_PKG = `${PKG_ROOT}/typings`;
 
 delete pkg.scripts;
 delete pkg.devDependencies;
 delete pkg.bundlesize;
 delete pkg.nyc;
+
+function copySources(rootDir, packageDir, ignoreMissing) {
+    // If we are ignoring missing directories, early return when source doesn't exist
+    if (!fs.existsSync(rootDir)) {
+        if (ignoreMissing) {
+            return;
+        } else {
+            throw "Source root dir does not exist!";
+        }
+    }
+    fs.copySync(rootDir, packageDir);
+}
 
 fs.removeSync(PKG_ROOT);
 
@@ -36,14 +46,13 @@ const rootPackageJson = {
 mkdirp.sync(PKG_ROOT);
 
 copySources(CJS_ROOT, CJS_PKG);
-cleanSourceMapRoot(PKG_ROOT, SRC_ROOT_PKG);
-
 fs.copySync(TYPE_ROOT, TYPE_PKG);
-
 copySources(ESM5_ROOT, ESM5_PKG, true);
-cleanSourceMapRoot(ESM5_PKG, SRC_ROOT_PKG);
-
 copySources(ESM2015_ROOT, ESM2015_PKG, true);
-cleanSourceMapRoot(ESM2015_PKG, SRC_ROOT_PKG);
+
+fs.copySync('./LICENSE.txt', `${PKG_ROOT}/LICENSE.txt`);
+fs.copySync('./README.md', `${PKG_ROOT}/README.md`);
 
 fs.writeJsonSync(`${PKG_ROOT}/package.json`, rootPackageJson, {spaces: 2});
+
+
