+++ package.json
@@ -54,7 +54,6 @@
     "expect.js": "~0.3.1",
     "fs-extra": "10.0.0",
     "github-api": "^3.4.0",
-    "klaw-sync": "6.0.0",
     "markdown-toc": "^1.2.0",
     "mkdirp": "^1.0.4",
     "mocha": "^8.0.1",
